# apps/desktop-companion/openapi.yaml
openapi: 3.1.0
info:
  title: FluxBoard Livestream Copilot API
  description: |
    REST API for the FluxBoard Livestream Copilot desktop companion service.

    FluxBoard provides real-time livestream capture, AI-powered content generation,
    OBS Studio integration, clip management, social media export, and video streaming
    for content creators.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Bearer JWT** - Pass an access token in the `Authorization` header.
    2. **API Key** - Pass an API key in the `x-api-key` header.

    Public endpoints (health checks, OBS status, event reads) do not require authentication
    during development.

    ## WebSocket Events

    The server exposes a WebSocket on a configurable port (default 9001). Clients receive
    real-time `EventEnvelope` messages for session lifecycle, transcripts, clips, outputs,
    and trigger detections. Connect and listen for JSON messages.

    Event types: `SESSION_START`, `SESSION_END`, `TRANSCRIPT_SEGMENT`, `MOMENT_MARKER`,
    `CLIP_INTENT_START`, `CLIP_INTENT_END`, `ARTIFACT_CLIP_CREATED`, `ARTIFACT_FRAME_CREATED`,
    `OUTPUT_CREATED`, `OUTPUT_VALIDATED`, `AUTO_TRIGGER_DETECTED`, `CLIP_QUEUE_UPDATED`.

    ## Rate Limiting

    Rate-limited endpoints return standard `RateLimit-*` headers. When exceeded, the server
    responds with `429 Too Many Requests`.

    | Endpoint Group        | Limit              |
...

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login or /api/v1/auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key for service-to-service communication
  schemas:
...
    SocialPlatform:
      type: string
      enum: [YOUTUBE, TWITTER, TIKTOK, LINKEDIN, INSTAGRAM, FACEBOOK, THREADS, BLUESKY]
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /api/v1/export/post:
    post:
      tags: [Export]
      summary: Export a social media post
      operationId: exportPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, platform]
              properties:
                text:
                  type: string
                  maxLength: 10000
                platform:
                  $ref: '#/components/schemas/SocialPlatform'
                sessionId:
                  type: string
                clipId:
                  type: string
                options:
                  type: object
                  properties:
                    copyToClipboard:
                      type: boolean
                    saveToFile:
                      type: boolean
                    optimizeHashtags:
                      type: boolean
                    addTimestamps:
                      type: boolean
                    createThread:
                      type: boolean
      responses:
        "201":
          description: Post exported
  /api/v1/export/clip:
    post:
      tags: [Export]
      summary: Export a video clip with format conversion
      operationId: exportClip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clipId, format]
              properties:
                clipId:
                  type: string
                format:
                  type: string
                  enum: [MP4, WEBM, GIF, MOV]
                platform:
                  $ref: '#/components/schemas/SocialPlatform'
                options:
                  type: object
                  properties:
                    quality:
                      type: string
                      enum: [low, medium, high, original]
                    generateThumbnail:
                      type: boolean
                    addWatermark:
                      type: boolean
                    optimizeForPlatform:
                      type: boolean
                    targetAspectRatio:
                      type: string
      responses:
        "201":
          description: Clip export started
  /api/v1/export/batch:
    post:
      tags: [Export]
      summary: Batch export multiple items
      operationId: exportBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [post, clip]
                      data:
                        type: object
                options:
                  type: object
                  properties:
                    zipOutput:
                      type: boolean
                    includeMetadata:
                      type: boolean
      responses:
        "201":
          description: Batch export started
  /api/v1/export/preview:
    post:
      tags: [Export]
      summary: Preview formatted post for multiple platforms
      operationId: previewPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, platforms]
              properties:
                text:
                  type: string
                  maxLength: 10000
                platforms:
                  type: array
                  items:
                    $ref: '#/components/schemas/SocialPlatform'
      responses:
        "200":
          description: Formatted previews per platform
  /api/v1/export/history:
    get:
      tags: [Export]
      summary: Get export history
      operationId: getExportHistory
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        "200":
          description: Export history list
...
  /api/v1/export/{id}/status:
    get:
      tags: [Export]
      summary: Get export job status
      operationId: getExportStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export status with progress
        "403":
          description: Unauthorized
        "404":
          description: Export not found
  /api/v1/export/{id}/download:
    get:
      tags: [Export]
      summary: Download export file
      operationId: downloadExport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File stream download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Export not ready
        "404":
          description: Export or file not found
  /api/v1/export/{id}:
    delete:
      tags: [Export]
      summary: Delete an export
      operationId: deleteExport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export deleted

  # ── Branding ─────────────────────────────────────────────────────────────
...
  /clip:
    post:
      tags: [Legacy]
      summary: Create a clip from replay buffer
      operationId: createClipLegacy
      security: []
      responses:
        "200":
          description: Clip created with artifact ID and paths
        "404":
          description: No active session
...